move `new_with_map_kind` on its own impl block
since this method needs the clock C to implement Default, and it's only
available for std environments, we better not force non-std clocks to
implement default for a method they don't even have access to.

ps. we might wanna re-consider whether we actually want to feature gate
`new_with_map_kind` to std only in the first place!